+++ server/project-service/src/main/java/edu/illinois/ncsa/incore/service/project/models/DFR3MappingResource.java
@@ -67,6 +67,30 @@ public class DFR3MappingResource extends ProjectResource {
         this.type = type;
     }
 
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setHazardType(String hazardType) {
+        this.hazardType = hazardType;
+    }
+
+    public String getHazardType() {
+        return hazardType;
+    }
+
+    public void setInventoryType(String inventoryType) {
+        this.inventoryType = inventoryType;
+    }
+
+    public String getInventoryType() {
+        return inventoryType;
+    }
+
     public boolean matchesSearchText(String text) {
         String lowerCaseText = text.toLowerCase();
         return (this.getId() != null && this.getId().equals(lowerCaseText)) ||