+++ server/project-service/src/main/java/edu/illinois/ncsa/incore/service/project/models/HazardResource.java
@@ -1,11 +1,7 @@
 package edu.illinois.ncsa.incore.service.project.models;
 
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import edu.illinois.ncsa.incore.common.data.models.jackson.JsonDateSerializer;
 
-import java.util.Date;
 import java.util.List;
 
 
@@ -24,9 +20,8 @@ public class HazardResource extends ProjectResource {
     public Type type = Type.earthquake;
 
     // Only keep basic information for now
-    public String name;
+    public String name = null;
     public String description;
-    public Date date = new Date();
     public String creator = null;
     public String owner = null;
     public List<HazardDataset> hazardDatasets;
@@ -43,9 +38,20 @@ public class HazardResource extends ProjectResource {
         this.type = type;
     }
 
-    @JsonSerialize(using = JsonDateSerializer.class)
-    public Date getDate() {
-        return date;
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setCreator(String creator) {
+        this.creator = creator;
+    }
+
+    public String getCreator() {
+        return creator;
     }
 
     public boolean matchesSearchText(String text) {