+++ server/project-service/src/main/java/edu/illinois/ncsa/incore/service/project/models/MappingEntryKey.java
@@ -10,10 +10,12 @@
 
 package edu.illinois.ncsa.incore.service.project.models;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import dev.morphia.annotations.Embedded;
 
 @Embedded
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class MappingEntryKey {
     public String name;
     public String description;