+++ server/space-service/src/main/java/edu/illinois/ncsa/incore/service/space/controllers/SpaceController.java
@@ -51,12 +51,12 @@ import java.util.stream.Collectors;
 @OpenAPIDefinition(
     info = @Info(
         description = "IN-CORE Space Service for creating and accessing spaces",
-        version = "1.27.1",
+        version = "1.28.0",
         title = "IN-CORE v2 Space Service API",
         contact = @Contact(
             name = "IN-CORE Dev Team",
             email = "incore-dev@lists.illinois.edu",
-            url = "https://incore.ncsa.illinois.edu"
+            url = "https://tools.in-core.org"
         ),
         license = @License(
             name = "Mozilla Public License 2.0 (MPL 2.0)",
@@ -75,7 +75,7 @@ import java.util.stream.Collectors;
 //        contact = @Contact(
 //            name = "IN-CORE Dev Team",
 //            email = "incore-dev@lists.illinois.edu",
-//            url = "https://incore.ncsa.illinois.edu"
+//            url = "https://tools.in-core.org"
 //        ),
 //        license = @License(
 //            name = "Mozilla Public License 2.0 (MPL 2.0)",